SQL-2919: integration test refactor
